[REFACTOR] TaskScheduler를 활용한 알림등록 클래스 이름 변경
기존: ViewingReminderScheduler.java
변경: TaskReminderScheduler.java

이유: 뷰잉에 한정된 task 등록하는 클래스가 아니기 때문에 범용적으로 이름 변경
