+++ src/main/java/org/hanihome/hanihomebe/verification/service/VerificationNotificationService.java
@@ -2,7 +2,7 @@ package org.hanihome.hanihomebe.verification.service;
 
 import lombok.RequiredArgsConstructor;
 import org.hanihome.hanihomebe.notification.application.service.NotificationFacadeService;
-import org.hanihome.hanihomebe.notification.application.service.NotificationMessageFactory;
+import org.hanihome.hanihomebe.notification.application.service.factory.NotificationMessageFactory;
 import org.hanihome.hanihomebe.notification.web.dto.NotificationCreateDTO;
 import org.springframework.stereotype.Service;
 @RequiredArgsConstructor
@@ -12,12 +12,12 @@ public class VerificationNotificationService {
     private final NotificationMessageFactory messageFactory;
 
     public void sendRejectNotification(Long verificationId, String reason) {
-        NotificationCreateDTO message = messageFactory.createVerificationRejectMessage(verificationId, reason);
+        NotificationCreateDTO message = messageFactory.getVerificationMessageCreator().createVerificationRejectMessage(verificationId, reason);
         notificationFacadeService.sendNotification(message);
     }
 
     public void sendApproveNotification(Long verificationId) {
-        NotificationCreateDTO message = messageFactory.createVerificationApproveMessage(verificationId);
+        NotificationCreateDTO message = messageFactory.getVerificationMessageCreator().createVerificationApproveMessage(verificationId);
         notificationFacadeService.sendNotification(message);
     }
 }
